nickserv/enforce: REGAIN: fix race condition
FNC semantics with most protocols are that services request the nick
change, but the actual nick change itself has to originate from the
user's own server. This means that we must wait for the Guest nick
change to be applied before changing the regaining user's nick, as
otherwise that change might be applied first on part of the network,
requiring an ircd-side resolution of the nick collision resulting from
this as the original person using the nick has not been renamed yet.

This probably still has a number of imperfections; it's essentially a
rebase of a handful of commits from almost two years ago and hasn't seen
much real-world testing yet.