chanfix: fix module deinit; make reload-only
As the chanfix module introduces DB entry types, it cannot be safely
unloaded.

The deinitialization code did not tear down various things that the
initialization code had set up. Additionally, in the specific case of
reloading the module, the initialization code failed to set up certain
things that had been torn down by the deinitialization code; as a
particularly obvious example, the deinit function would delete the
chanfix service itself but the init function would not re-add it if the
module was being reloaded.

This commit hopefully addresses these issues. It also makes the deinit
code unconditionally assume the module is being reloaded as the module
is set to be reload-only anyway.

Due to several issues with the previous deinitialization code, it is
strongly recommended you avoid reloading chanfix/main if you are
currently using the previous version; instead, you should restart
services with the updated version of chanfix/main.