misc/canon_domains: new email canonicalizer module
This module is mainly intended to handle email providers with multiple
interchangable domains serving the same users on each.

For now this is rather simplistic, requiring mappings to be defined in
the configuration file and not supporting wildcards or anything fancier
than "treat these domains as the same".

All domains of a given provider will be mapped to the shortest domain to
avoid having to handle the case of extending the length of an email
address, possibly exceeding EMAILLEN.

(Do note that nickserv/badmail currently does not interface with email
canonicalization and domains need to be added to that list separately
if they are to be denied.)