chanserv/list: ignore fantasy invocations
This command isn't channel-specific, so there isn't really any reason to
use it via a fantasy command. Meanwhile, this allows bots to respond to
the command instead (cf chanserv/register doing the same, although that
one manages to be even *more* pointless for fantasy use).

(cf comment about this on #721)