modules/operserv/modmanager.c: reduce recursion stack impact
We're already using a field width specifier to align the module name
(relative to the length of the largest name), so use a field width
specifier to align the depth (indentation) of the module name too, by
asking it to align the empty string.

This saves 1k of stack for every recursive call.