Compiler Sanitizers: Require that the heap allocator is disabled
It makes little sense to enable the various sanitizers (like ASAN
and UBSAN) if the shared heap allocator feature is enabled, as that
can mask various memory errors like use-after-free, which defeats
their entire purpose.

Also clarify in README.md that Clang is not required to use the
sanitizers, but it is recommended; and that sanitizers are not
recommended for production usage (by end-users).

Finally unify the format of the error messages for conflicting
configure arguments with that of the sodium memory allocator test.