Compiler Sanitizers: Require that the Sodium memory allocator is disabled
Same rationale as for the previous commit. Anything that interferes with
how memory is allocated, initialised, and freed, will have a detrimental
effect on the compiler sanitizers, especially ASAN.

The documentation is not modified because the Sodium memory allocator is
not enabled by default, unlike the shared heap allocator and linker library
stripping.