libathemecore/logger.c: vslog_ext(): tidy up & fix log suppression bug
The previous code looped over the configured log files and returned from
the function if one of them did not have a write handler. However, this
would prevent the static variable "in_slog" from being reset back to
false, suppressing all further log messages. [1]

Also, rename this variable to reflect the name of the function it is in,
because slog_ext() is called from more than just slog() [in the
logcommand() family of functions and in logaudit_denycmd()].

Finally, tidy up the code a little; moving variable declarations to
where they are needed, not performing computation outside of a
conditional block if the results are only used inside of it, not
performing explicit casts from (void *), etc.

[1] This is used to detect infinite logging recursion. That could occur,
    for example, if services is configured to log raw network traffic to
    an IRC channel. We don't want to log that we're sending a log line,
    and then log that we're sending that log line, and so on...