./configure: hard-fail if compiler is not C99-compliant
We use a lot of C99 features in this codebase, so a compiler that is not
C99-compliant is unlikely to be able to build this codebase successfully.

Below, [X] denotes a C99 feature that this codebase currently uses,
excluding the C99 compliant snprintf(3) fallback implementation in
libathemecore (which uses va_copy too), as that is due to be removed for
the same reason; we require a C99 compiler, so we should not need a C99
compliant snprintf(3) fallback implementation at all.

https://www.gnu.org/software/autoconf/manual/autoconf-2.69/html_node/C-Compiler.html

    Macro: AC_PROG_CC_C99

    If the C compiler is not in C99 mode by default, try to add an option
    to output variable CC to make it so. This macro tries various options
    that select C99 on some system or another, preferring extended
    functionality modes over strict conformance modes.

    It considers the compiler to be in C99 mode if it handles:

        [X] _Bool
        [X] // comments
            flexible array members
        [X] inline
        [X] signed and unsigned long long int
        [X] mixed code and declarations
        [X] named initialization of structs
        [X] restrict
            va_copy
        [X] varargs macros
        [X] variable declarations in for loops
        [X] variable length arrays

    After calling this macro you can check whether the C compiler has been
    set to accept C99; if not, the shell variable ac_cv_prog_cc_c99 is set
    to 'no'.