Direct Digest API: Ensure SHA2 state buffers are suitably aligned
This was already the case; as there is no padding required between the
members, the largest member has an alignment of 8 bytes (thus ensuring
that the struct is also aligned to 8 bytes), and the sizes of all
members before the state buffer combine to a multiple of 8 bytes anyway.

However, since the code does cast this buffer to an unsigned 64-bit
integer, make sure that it is aligned to a 64-bit boundary, for processor
architectures which require aligned reads.

This will ensure that the state buffer is suitably aligned even if the
compiler decides to, for some completely unjustifiable reason, insert
padding before it.

Then we can silence the absolutely pointless -Wcast-align diagnostic for
this cast, by performing an intermediate cast to (void *), which has no
size, and thus no alignment. That code could have been rewritten to use
memcpy(3), and then alignment wouldn't have been required either, but I
think that would end up looking a bit weird.