Remove the libsodium hardened memory allocator
This has served its purpose, and with the maturity of various compiler
sanitizers (and AddressSanitizer in particular) in gcc and clang, this
has little use over just providing "--enable-compiler-sanitizers" to
./configure, and will catch much fewer memory errors.

The sodium memory allocator returns a pointer (PAGESIZE-num) bytes into
a page (for an allocation of num bytes, assuming num is not greater than
PAGESIZE), ensuring that a 1-byte overrun will terminate the process
(because it puts a no-access guard page immediately after it). It also
puts a no-access guard page before the allocated page(s), preventing an
unrelated overrun from having an undesired effect. Finally, it fills the
portion of the page that isn't being used for the user allocation with a
canary value, which is checked on sodium_free() to detect under-writes.

However, AddressSanitizer can do much more than this, without as much
virtual memory overhead per allocation. It can *immediately* detect
(and abort on) a 1+-byte under-write (as opposed to when the memory is
freed), and can *immediately* detect (and abort on) a 1+-byte under-
read, which isn't possible with the sodium allocator at all (unless the
application under-reads so much that it hits the preceeding guard page).
Finally, it can provide a stack backtrace and a memory map of the fault,
and also describe which variable was overrun and which other variables
are nearby, enabling the cause of the abort to be detected more easily.