infoserv/main: code deduplication
The existing implementation had most of its code existing twice, one
version handling normal logon info and the other version handling
oper-logon info.

This replaces most of the duplicated code with a single piece each that
handles both types of entry via the same codepath. In particular,
struct oper_logoninfo was identical to struct logoninfo; it has thus
been replaced with the latter.

Command handlers are replaced with a stub handler that calls into the
actual implementation with an added parameter to indicate which list to
operate on.

Some of the code was slightly cleaned up as well:
- minor indentation fix
- move variable declarations to where they are needed
- use #define for DB type identifiers