infoserv/main: add logoninfo_reverse option
Previously, InfoServ would show messages in a "newest first" order. This
allows that behaviour to be reversed, showing messages in chronological
order instead.

This interacts with logoninfo_count to make any messages added after
that limit not display at connection time. Warn about this in the
configuration and whenever adding a message that will be affected.