modules/scripting/perl: make non-unloadable
Reloading this module causes services to segfault because the
PERL_SYS_INIT3() macro should only be called once during the entire
lifetime of the process [1]. Allowing it to be unloaded thus carries the
risk that it will be reloaded (or, at a later date, loaded again).

I tried to (ab)use Mowgli's global storage API to only call it once
(regardless of how many times it's loaded), and to not call the
PERL_SYS_TERM() macro at all, but this only lead to a different crash on
reload: trying to allocate a little less than 100 TiB of memory (!).

If a Perl expert comes along and weighs in, this commit can be reverted
and the underlying problem fixed. Nonetheless, libperl did exhibit *dozens*
of uses of unintialized data on reload (confirmed by valgrind), followed by
a segmentation fault even if we skip calling PERL_SYS_INIT3() again
(because, naturally, such a large allocation will usually return NULL, and
it apparently doesn't deal with that).

[1] https://perldoc.perl.org/perlembed