libathemecore/services.c: get_source_name(): complete NULL checks
Code above this already tests if si->v is NULL; indicating that it can
be NULL (and it can be). Therefore, check if it's NULL again before
attempting to dereference it again.

At present this cannot be the case, but guard against it anyway, making
sure to fill the result buffer with something in any case.

Identified by Coverity Scan.