libathemecore/logger.c: logfile_new(): fix up close-on-exec logic
We were calling fcntl(2) F_SETFD without first obtaining the current
file descriptor flags with F_GETFD. Furthermore, we were not checking
the return value of the F_SETFD operation. Now we obtain the current
flags to bitwise-OR FD_CLOEXEC them with, and we warn if the operation
fails.

However, it would be better if we didn't have to try to call it in the
first place, so try opening the file descriptor with the O_CLOEXEC flag
first, which is safer too.

While we're at it, make sure that the log file doesn't end up being
world-readable if services' umask is not sufficient to prevent this.

Identified by Coverity Scan.