Build System: Detect user error w/ downloading sourcecode
A frequent complaint we receive is that serno.h is missing. This turns
out to be users downloading release tarballs from GitHub, which does
not include the .git directory, thus making the target for serno.h
fail.

Detect whether .git/ exists; reference GIT-Access.txt then. Otherwise,
this must be a release tarball, so see if it has a pre-supplied serno.h
and error if not (this could only be caused by downloading a GitHub
source code link instead of an asset/release tarball).

Also tidy up mkserno.sh a bit. It still can't use git-describe(1) on
this branch (lack of recent tags), but will be made to do so on release
branches.

Thanks to ilbelkyr for the idea (pointing out that the sourcecode links
on GitHub don't include the .git/ directory, making it possible to
distinguish between that and git-clone(1)).