configure.ac: prepend LIBPCRE_CFLAGS and LIBPCRE_LIBS
The entire codebase depends on libpcre if it is detected successfully,
so we need to build all of the .c files with the appropriate CFLAGS &
LIBS.

This was taken care of in libathemecore/Makefile, but nowhere else.
It's better to just have configure.ac take care of adding it if they
are non-empty.