m4/atheme-api-random-frontend.m4: detect more secure arc4random(3)s
It turns out that a lot more than recent OpenBSD systems have a
secure algorithm backing their arc4random(3) implementation.

Remove the check for just OpenBSD, and add checks for:

- Mac OS 10.12+
- FreeBSD 12.0+
- NetBSD 7.0+
- OpenBSD 5.4+

Created with input from rqsd from IRC, in response to commit
8ff53e4d2a843c5fdd7a.

Note that this is not a statement that Atheme will successfully
compile & run on the above platforms.