libathemecore/connection.c: use connection->name to store addrs & ports
Also obtain IP addresses where the sockets are created, instead of in
connection_add(). This is a better approach than that taken by commit
1bb7e1e587306239ca87, and restores the ability to perform non-blocking
connections.

Finally, since connection_add() creates the mowgli.pollable object for
the fd, don't bother testing if the fd is valid. Add an assertion for a
valid fd being passed, because an invalid fd will only result in
nothing working anyway. Likewise assert that the connection name is
valid, and that at least one I/O handler was supplied.