Add match_masks_through_vhost option for Solanum
This reflects a similar Solanum PR: solanum-ircd/solanum#150

It might be slightly more proper to have this option apply specifically
to the solanum protocol module once we have one, but we don't, plus
there isn't currently any precedent for config options added from
protocol modules.

The option is deliberately rehashable as the Solanum option is as well,
allowing a synchronized change in configuration across a network without
requiring restarts.