libathemecore/ptasks.c: handle_kill(): don't use slog() if we got killed
If a client on our server is killed, we cannot reliably use slog(),
because the client may be OperServ, and slog() mesages are sent from
OperServ.

It turns out that some IRCds care about receiving messages from UIDs
it doesn't know about (because it just killed them), and will send
KILLs for them over and over again.

Reported-By: @A-UNDERSCORE-D