email/default/setpass: don't encourage contact for unsolicited receipt
Nothing can be done to the account without the token in this e-mail,
and any competent malicious actor who does have unauthorised access to
a user's e-mail inbox is also just going to delete the e-mail after
compromising the user's account anyway.

Therefore, it doesn't make sense to encourage the user to reach out to
the network administration.

Also clarify that password reset tokens are now automatically
invalidated by a successful login.