libathemecore/function.c: sendemail(): allow for urlencoding account
If one wishes to replace their register email template to prefix the account
name and registration verification token with a URL (to implement a webserver
to verify registrations, so people don't have to copy and paste commands to
execute on IRC), we should allow for the possibility that people will have
non-alphanumeric characters in their account name, and provide a URL-encoded
version of it.

This will ensure that such links are always clickable in various MUAs.