libathemecore/ptasks: handle status-prefixed channel messages
If a channel message comes in with a status prefix, we should still
treat it as a channel message.

This does not raise concerns about chanops hiding the usage of fantasy
commands because they can always private message services directly and
the output of potentially-abusive commands (for example, kick) is
signed off with their account name anyway (which will not change).

Under the solanum family of IRCds, there exists a channel mode +z,
which will transform messages to #channel into messages to @#channel
if the message would have been blocked by +m/+b/+q. This is to clearly
distinguish to channel operators that the only reason they are seeing
the message is that they are operators, and that other channel users
will not have seen it.

ChanServ is also a channel operator, but it would not have received
the message, because services ignored status-prefixed channel messages.
This prevented users using fantasy commands if they are affected by +z;
for example, by the channel being +m. Confusingly, this would have
allowed "!deop" to work, but not "!op".

Reported by juggs on IRC, who did not respond to a request for the
proper attribution in Reported-by.