Hook: user_can_login: Indicate login method
This allows hooks to permit or deny logins based upon the type
of credential being used.

This requires some rework of how SASLServ behaves. Specifically,
mechanism modules now indicate their login type at the point
where the hook is called, not when the mechanism is registered.

At the moment, there are four types of login credential:

- Certificate Fingerprints
  - User introduction (not logged in, but has a certfp)
  - SASL EXTERNAL

- Passwords
  - NickServ IDENTIFY
  - SASL PLAIN
  - SASL SCRAM

- Public-key Challenges
  - SASL ECDSA-NIST256P-CHALLENGE
  - SASL ECDH-X25519-CHALLENGE

- Tokens
  - SASL AUTHCOOKIE