modules/saslserv/main.c: sasl_handle_login(): fix overzealous message
This can be triggered simply by a client beginning a pre-registration
SASL login, and then completing registration before completing the
login (for example, a timeout waiting for the server to respond to our
AUTHENTICATE command which then results in a CAP END).

As such, the log level and text of this message is wrong.