Several minor cleanups and tweaks to login throttling
Choose sane defaults for both throttling mechanisms and actually have
them be separate so that the configuration options for one of them do
not render the other one impossible to trigger.

Document the fact that the IP address check is performed first.

Convert the configuration parsing code from using `float` to `double`.
During concept, I advised `float` under the assumption that it would
have enough range and precision to account for our specific needs.
However, during testing, the fact that this was not true became
apparent, and this does make it a bit more future-proof too.

Add missing declaration for new configuration parsing function.

Pass a double to check functions instead of a time_t that has to be
casted to double everywhere.

Check that the user is on IRC before checking if the user is an IRC
Operator.

Remove call to list_unregister() (and associated header inclusion)
for something that is now no longer passed to list_register().

Other misc coding style touchups and deduplication of a string.