ChanServ HELP FLAGS: Correct description of flag +r
What this flag allows you to do varies depending upon what combinations
of the following modules are loaded:

- chanserv/ban
- chanserv/kick
- chanserv/quiet
- chanserv/unban_self

Note that it is not possible to load both chanserv/ban and
chanserv/unban_self. Both of them provide an UNBAN command,
while only the former provides a BAN command, which would be
erroneous to describe if you have the latter loaded instead.