modules/nickserv/multimark: MARK ADD: check operator is logged in
The ADD subcommand takes a copy of the operator's account name and
entity ID for future LIST output.

Ensure that the operator is actually logged into a services account
before attempting to do this, or we will crash.

It is only possible to trigger this crash by giving MARK privileges
to the "ircop" operclass, being opered, and not being logged in
while executing MARK ADD.

Reported-By: TheDaemoness <she@daemoness.dev>