chanserv/set_mlock: improve logging for oper mlock
When privileged uesrs adjust oper-only modes in a channel's mlock
setting, we used to generate a second CMDLOG_SET line that was hard to
distinguish from the regular MLOCK log line and didn't add much in the
way of useful information. This commit logs the specific changes made
(specifically to oper-only modes) to CMDLOG_ADMIN instead.

As changing oper-only modes in MLOCK can be done regardless of channel
access and may have rather significant effects on the channel under
consideration (e.g. Solanum-style +M preventing opers from being
kicked), this seems to warrant erring on the side of caution and logging
to wallops() as well.

Fixes #832.